+++ test/ApplicationTest.php
@@ -68,6 +68,16 @@ class ApplicationTest extends TestCase
         $this->assertSame($this->noopMiddleware, $route->getMiddleware());
     }
 
+    public function testAnyRouteMethod()
+    {
+        $this->router->addRoute(Argument::type(Route::class))->shouldBeCalled();
+        $route = $this->getApp()->any('/foo', $this->noopMiddleware);
+        $this->assertInstanceOf(Route::class, $route);
+        $this->assertEquals('/foo', $route->getPath());
+        $this->assertSame($this->noopMiddleware, $route->getMiddleware());
+        $this->assertSame(Route::HTTP_METHOD_ANY, $route->getAllowedMethods());
+    }
+
     /**
      * @dataProvider commonHttpMethods
      */